+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 ## main
 
 - REMOVED: Removed deprecated `get_domain_premium_price`. Use `get_domain_prices` instead.
+- REMOVED: Removed deprecated `get_whois_privacy` (dnsimple/dnsimple-developer#919)
+- REMOVED: Removed deprecated `renew_whois_privacy` (dnsimple/dnsimple-developer#919)
 
 ## 6.1.0