+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 7.1.0 - 2026-02-26
+
+### Added
+
+- Added `get_domain_research_status` to research a domain for availability and registration status. (#486)
+
 ## 7.0.0 - 2026-01-23
 
 ### Removed